% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd_lm.R
\name{sgd_lm}
\alias{sgd_lm}
\title{Averaged SGD and its Inference via Random Scaling}
\usage{
sgd_lm(
  x,
  y,
  gamma_0 = 1,
  alpha = 0.667,
  burn = 1,
  bt_start = NULL,
  path_output = NULL,
  studentize = TRUE,
  intercept = TRUE
)
}
\arguments{
\item{x}{numeric. (n x p) matrix of regressors. Should not include 1 (the intercept)}

\item{y}{numeric}

\item{gamma_0}{numeric}

\item{alpha}{numeric}

\item{burn}{numeric}

\item{bt_start}{numeric}

\item{path_output}{numeric specifying the sequence that print out the output paths}

\item{studentize}{logical. Studentize regressors. Default is TRUE}

\item{intercept}{logical. Use the intercept term for regressors. Default is TRUE}
}
\value{
#' An object of class \code{"sgd_lm"}, which is a list containing the following
components:
}
\description{
Compute the averaged SGD estimator and the confidence intervals via random scaling method.
}
\examples{
n = 1e05
p = 5
bt0 = rep(5,p)
x = matrix(rnorm(n*(p-1)), n, (p-1))
y = cbind(1,x) \%*\% bt0 + rnorm(n)
sgd_lm.out = sgd_lm(x,y)
}
