% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sgd_lm_cpp}
\alias{sgd_lm_cpp}
\title{Leading NA}
\usage{
sgd_lm_cpp(x, y, burn, gamma_0, alpha, bt_start)
}
\arguments{
\item{x}{numeric. (n x p) matrix of regressors. Should not include 1 (the intercept)}

\item{y}{numeric}

\item{burn}{numeric}

\item{gamma_0}{numeric}

\item{alpha}{numeric}

\item{bt_start}{numeric}
}
\description{
Explanation
}
