% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd_sqr.R
\name{sgd_sqr}
\alias{sgd_sqr}
\title{Averaged SGD and its Inference via Random Scaling}
\usage{
sgd_sqr(
  x,
  y,
  gamma_0 = 1,
  alpha = 0.667,
  burn = 1,
  bt_start = NULL,
  path_output = NULL,
  qt = 0.5,
  studentize = TRUE,
  intercept = TRUE,
  h = 0
)
}
\arguments{
\item{x}{numeric. (n x p) matrix of regressors. Should not include 1 (the intercept)}

\item{y}{numeric}

\item{gamma_0}{numeric}

\item{alpha}{numeric}

\item{burn}{numeric}

\item{bt_start}{numeric}

\item{path_output}{numeric specifying the sequence that print out the output paths}

\item{qt}{numeric. Quantile. Default is 0.5.}

\item{studentize}{logical. Studentize regressors. Default is TRUE}

\item{intercept}{logical. Use the intercept term for regressors. Default is TRUE}

\item{h}{numeric. Bandwidth for smoothed QR estimator. Default is 0 (automatic choice)}
}
\value{
#' An object of class \code{"sgdi"}, which is a list containing the following
components:
}
\description{
Compute the averaged SGD estimator and the confidence intervals via random scaling method.
}
