% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdi.R
\name{sgdi}
\alias{sgdi}
\title{Averaged SGD and its Inference via Random Scaling}
\usage{
sgdi(
  x,
  y,
  gamma_0 = 1,
  alpha = 0.667,
  burn = 1,
  model = "lm",
  inference = "rs",
  bt_start = NULL,
  path_output = NULL,
  qt = 0.5,
  studentize = TRUE,
  intercept = TRUE,
  rss_idx = c(1)
)
}
\arguments{
\item{x}{numeric. (n x p) matrix of regressors. Should not include 1 (the intercept)}

\item{y}{numeric. (n x 1) vector of a dependent variable.}

\item{gamma_0}{numeric. A tuning parameter for the learning rate (gamma_0 x t ^ alpha). Default is 1.}

\item{alpha}{numeric. A tuning parameter for the learning rate (gamma_0 x t ^ alpha). Default is 0.667.}

\item{burn}{numeric. A tuning parameter for "burn-in" observations. We burn-in up to (burn-1) observations and use observations from (burn) for estimation. Default is 1, i.e. no burn-in.}

\item{model}{character specifying the model to be used: \code{"lm"} (linear
mean regression model), \code{"qr"} (quantile regression)}

\item{inference}{character. Specifying the inference method. Default is "rs" (random scaling). "rss" is for ransom scaling subset inference. Then, "rss_indx" should be provided.}

\item{bt_start}{numeric. (p x 1) vector. User-provided starting value Default is NULL.}

\item{path_output}{numeric specifying the sequence that print out the output paths}

\item{qt}{numeric. Quantile. Default is 0.5.}

\item{studentize}{logical. Studentize regressors. Default is TRUE}

\item{intercept}{logical. Use the intercept term for regressors. Default is TRUE}

\item{rss_idx}{numeric. Index of x for random scaling subset inference. Default is 1, the first regressor of x. For example, if we want to infer the 1st, 3rd covariate of x, then set it to be c(1,3).}
}
\value{
An object of class \code{"sgdi"}, which is a list containing the following
\describe{
\item{\code{beta_hat}}{A (p + 1)-vector of estimated parameter values including the intercept.}
\item{\code{V_hat}}{A (p+1)x (p+1) variance-covariance matrix of \code{beta_hat}}
\item{\code{V_hat_sub}}{A variance-covariance sub-matrix of \code{beta_hat}. If the subset size is not provided, it returns 0.}
}
}
\description{
Compute the averaged SGD estimator and the variance-covariance matrix via random scaling method.
}
\examples{
n = 1e05
p = 5
bt0 = rep(5,p)
x = matrix(rnorm(n*(p-1)), n, (p-1))
y = cbind(1,x) \%*\% bt0 + rnorm(n)
out = sgdi(x,y)
}
