% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdi_boot_qr.R
\name{sgdi_boot_qr}
\alias{sgdi_boot_qr}
\title{Inference of quantile regression with SGD and random scaling}
\usage{
sgdi_boot_qr(
  x,
  y,
  gamma_0 = 1,
  alpha = 0.667,
  burn = 1,
  inference = "boot",
  bt_start = NULL,
  path_output = NULL,
  qt = 0.5,
  studentize = TRUE,
  intercept = TRUE,
  n_boot = 1000
)
}
\arguments{
\item{x}{numeric. (n x p) matrix of regressors. Should not include 1 (the intercept)}

\item{y}{numeric}

\item{gamma_0}{numeric. Tuning parameter for the learning rate: gamma_0 * n^(alpha) (default is 1)}

\item{alpha}{numeric. Tuning parameter for the learning rate: gamma_0 * n^(alpha) (default is 0.667)}

\item{burn}{numeric. Number of burn-in observations (default is 1, no burn-in)}

\item{inference}{character specifying the inference method. Default is "rs" (random scaling)}

\item{bt_start}{numeric.}

\item{path_output}{numeric specifying the sequence that print out the output paths}

\item{qt}{numeric. Quantile. Default is 0.5.}

\item{studentize}{logical. Studentize regressors. Default is TRUE}

\item{intercept}{logical. Use the intercept term for regressors. Default is TRUE}
}
\value{
#' An object of class \code{"sgdi"}, which is a list containing the following
components:
}
\description{
Compute the averaged SGD estimator of quantile regression and the confidence intervals via random scaling method.
}
\examples{
n = 1e05
p = 5
bt0 = rep(5,p)
x = matrix(rnorm(n*(p-1)), n, (p-1))
y = cbind(1,x) \%*\% bt0 + rnorm(n)
sgdi.out = sgdi_qr(x,y)
}
