% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sgdi_lm_cpp}
\alias{sgdi_lm_cpp}
\title{Leading NA}
\usage{
sgdi_lm_cpp(x, y, burn, gamma_0, alpha, bt_start, inference)
}
\arguments{
\item{x}{numeric. (n x p) matrix of regressors. Should not include 1 (the intercept)}

\item{y}{numeric}

\item{burn}{numeric}

\item{gamma_0}{numeric}

\item{alpha}{numeric}

\item{bt_start}{numeric}

\item{inference}{character specifying the inference method. Default is "rs" (random scaling)}
}
\description{
Explanation
}
