% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdi_qr.R
\name{sgdi_qr}
\alias{sgdi_qr}
\title{Inference of quantile regression with SGD and random scaling}
\usage{
sgdi_qr(
  formula,
  data,
  gamma_0 = 1,
  alpha = 0.667,
  burn = 1,
  inference = "rs",
  bt_start = NULL,
  qt = 0.5,
  studentize = TRUE,
  intercept = TRUE,
  rss_idx = c(1)
)
}
\arguments{
\item{formula}{formula. The response is on the left of a ~ operator. The terms are on the right of a ~ operator, separated by a + operator.}

\item{data}{an optional data frame containing variables in the model.}

\item{gamma_0}{numeric. A tuning parameter for the learning rate (gamma_0 x t ^ alpha). Default is 1.}

\item{alpha}{numeric. A tuning parameter for the learning rate (gamma_0 x t ^ alpha). Default is 0.667.}

\item{burn}{numeric. A tuning parameter for "burn-in" observations. We burn-in up to (burn-1) observations and use observations from (burn) for estimation. Default is 1, i.e. no burn-in.}

\item{inference}{character. Specifying the inference method. Default is "rs" (random scaling). "rss" is for ransom scaling subset inference. Then, "rss_indx" should be provided.}

\item{bt_start}{numeric. (p x 1) vector. User-provided starting value Default is NULL.}

\item{qt}{numeric. Quantile. Default is 0.5.}

\item{studentize}{logical. Studentize regressors. Default is TRUE}

\item{intercept}{logical. Use the intercept term for regressors. Default is TRUE}

\item{rss_idx}{numeric. Index of x for random scaling subset inference. Default is 1, the first regressor of x. For example, if we want to infer the 1st, 3rd covariate of x, then set it to be c(1,3).}
}
\value{
An object of class \code{"sgdi"}, which is a list containing the following
\describe{
\item{\code{beta_hat}}{A (p + 1)-vector of estimated parameter values including the intercept.}
\item{\code{V_hat}}{A (p+1)x (p+1) variance-covariance matrix of \code{beta_hat}}
\item{\code{V_hat_sub}}{A variance-covariance sub-matrix of \code{beta_hat}. If the subset size is not provided, it returns 0.}
}
}
\description{
Compute the averaged SGD estimator of quantile regression and the confidence intervals of quantile regressionvia random scaling method.
}
\examples{
n = 1e05
p = 5
bt0 = rep(5,p)
x = matrix(rnorm(n*(p-1)), n, (p-1))
y = cbind(1,x) \%*\% bt0 + rnorm(n)
my.dat = data.frame(y=y, x=x)
sgdi.out = sgdi_qr(y~., data=my.dat)
}
