\name{SGDinference-package}
\alias{SGDinference-package}
\alias{SGDinference}
\docType{package}
\title{
\packageTitle{SGDinference}
}
\description{
\packageDescription{SGDinference}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SGDinference}
\packageIndices{SGDinference}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SGDinference}

Maintainer: \packageMaintainer{SGDinference}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
