% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-init.R
\docType{package}
\name{SGDinference}
\alias{SGDinference}
\title{SGDinference}
\description{
ADD DESCRIPTION LATER
}
\author{
Sokbae (Simon) Lee <sl3841@columbia.edu>, Yuan Liao <yuan.liao@rutgers.edu>, Myung Hwan (Matt) Seo <myunghseo@snu.ac.kr>, Youngki Shin <shiny11@mcmaster.ca>
}
