% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd_qr.R
\name{sgd_qr}
\alias{sgd_qr}
\title{Averaged SsubGD estimator for quantile regression}
\usage{
sgd_qr(
  formula,
  data,
  gamma_0 = 1,
  alpha = 0.667,
  burn = 1,
  bt_start = NULL,
  qt = 0.5,
  studentize = TRUE,
  intercept = TRUE
)
}
\arguments{
\item{formula}{formula. The response is on the left of a ~ operator. The terms are on the right of a ~ operator, separated by a + operator.}

\item{data}{an optional data frame containing variables in the model.}

\item{gamma_0}{numeric. A tuning parameter for the learning rate (gamma_0 x t ^ alpha). Default is 1.}

\item{alpha}{numeric. A tuning parameter for the learning rate (gamma_0 x t ^ alpha). Default is 0.667.}

\item{burn}{numeric. A tuning parameter for "burn-in" observations. We burn-in up to (burn-1) observations and use observations from (burn) for estimation. Default is 1, i.e. no burn-in.}

\item{bt_start}{numeric. (p x 1) vector. User-provided starting value Default is NULL.}

\item{qt}{numeric. Quantile. Default is 0.5.}

\item{studentize}{logical. Studentize regressors. Default is TRUE}

\item{intercept}{logical. Use the intercept term for regressors. Default is TRUE}
}
\value{
An object of class \code{"sgd_qr"}, which is a list containing the following
\describe{
\item{\code{beta_hat}}{A (p + 1)-vector of estimated parameter values including the intercept.}
}
}
\description{
Compute the averaged SGD estimator for quantile regression. This function computes only the point estimate.
}
\examples{
n = 1e05
p = 5
bt0 = rep(5,p)
x = matrix(rnorm(n*(p-1)), n, (p-1))
y = cbind(1,x) \%*\% bt0 + rnorm(n)
my.dat = data.frame(y=y, x=x)
sgd.out = sgd_qr(y~., data=my.dat)
}
