% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDinference-package.R
\docType{package}
\name{SGDinference-package}
\alias{SGDinference}
\alias{SGDinference-package}
\title{SGDinference: Inference with Stochastic (sub-)Gradient Descent}
\description{
The package provides estimation and inference methods for large-scale mean and quantile regression models via stochastic (sub-)gradient descent (S-subGD) algorithms. The inference procedure handles cross-sectional data sequentially: (i) updating the parameter estimate with each incoming "new observation", (ii) aggregating it as a Polyak-Ruppert average, and (iii) computing an asympotically pivotal statistic for inference through random scaling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SGDinference-Lab/SGDinference/}
  \item Report bugs at \url{https://github.com/SGDinference-Lab/SGDinference/issues}
}

}
\author{
\strong{Maintainer}: Youngki Shin \email{shiny11@mcmaster.ca}

Authors:
\itemize{
  \item Sokbae Lee \email{sl3841@columbia.edu}
  \item Yuan Liao \email{yuan.liao@rutgers.edu}
  \item Myung Hwan Seo \email{myunghseo@snu.ac.kr}
}

}
\keyword{internal}
