% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDinference.R
\docType{package}
\name{SGDinference}
\alias{SGDinference}
\title{SGDinference}
\description{
The SGDinference package provides estimation and inference methods for large-scale mean and quantile regression models via stochastic (sub-)gradient descent (S-subGD) algorithms. 
The inference procedure handles cross-sectional data sequentially: 
(i) updating the parameter estimate with each incoming "new observation", 
(ii) aggregating it as a Polyak-Ruppert average, and 
(iii) computing an asymptotically pivotal statistic for inference through random scaling.
}
\author{
Sokbae Lee, Yuan Liao, Myung Hwan Seo, Youngki Shin
}
