% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-init.R
\docType{package}
\name{SGDinference}
\alias{SGDinference}
\title{SGDinference}
\description{
The SGDinference package comptues the averaged SGD estimator and the variance-covariance matrix via random scaling method.
}
\author{
SGDinference-Lab
}
